решение задачи 3

+++ Семинар 1/Задача 3/Program.cs
@@ -1,9 +1,9 @@
 ﻿// С клавиатуры вводятся два числа a и b. Найти максимальное из них. 
 Console.WriteLine("Введите первое число");
-double a = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToDouble(Console.ReadLine());
 
 Console.WriteLine("Введите второе число");
-double b = Convert.ToInt32(Console.ReadLine());
+double b = Convert.ToDouble(Console.ReadLine());
 
 
 if (b > a)
